privacy/secure_aggregation.py
"""Secure Aggregation for Federated Learning"""

import torch
import numpy as np
from typing import List, Dict, Tuple
from Crypto.PublicKey import RSA
from Crypto.Cipher import PKCS1_OAEP
import pickle


class SecureAggregator:
    """
    Secure aggregation using simple encryption.
    
    Args:
        num_clients: Number of clients
        key_size: RSA key size in bits
    """
    
    def __init__(self, num_clients: int, key_size: int = 2048):
        self.num_clients = num_clients
        self.key_size = key_size
        
        # Generate server key pair
        self.server_key = RSA.generate(key_size)
        self.public_key = self.server_key.publickey()
        
        print(f"Secure Aggregator initialized for {num_clients} clients")
    
    def get_public_key(self) -> RSA.RsaKey:
        """
        Get server public key for clients.
        
        Returns:
            Public key
        """
        return self.public_key
    
    def encrypt_weights(self, weights: List[np.ndarray]) -> bytes:
        """
        Encrypt model weights.
        
        Args:
            weights: List of weight arrays
            
        Returns:
            Encrypted weights as bytes
        """
        # Serialize weights
        weights_bytes = pickle.dumps(weights)
        
        # Encrypt using public key
        cipher = PKCS1_OAEP.new(self.public_key)
        
        # Split into chunks (RSA has size limits)
        chunk_size = self.key_size // 8 - 42  # OAEP padding overhead
        encrypted_chunks = []
        
        for i in range(0, len(weights_bytes), chunk_size):
            chunk = weights_bytes[i:i + chunk_size]
            encrypted_chunk = cipher.encrypt(chunk)
            encrypted_chunks.append(encrypted_chunk)
        
        return pickle.dumps(encrypted_chunks)
    
    def decrypt_weights(self, encrypted_weights: bytes) -> List[np.ndarray]:
        """
        Decrypt model weights.
        
        Args:
            encrypted_weights: Encrypted weights as bytes
            
        Returns:
            List of decrypted weight arrays
        """
        # Deserialize encrypted chunks
        encrypted_chunks = pickle.loads(encrypted_weights)
        
        # Decrypt using private key
        cipher = PKCS1_OAEP.new(self.server_key)
        
        decrypted_bytes = b''
        for encrypted_chunk in encrypted_chunks:
            decrypted_chunk = cipher.decrypt(encrypted_chunk)
            decrypted_bytes += decrypted_chunk
        
        # Deserialize weights
        weights = pickle.loads(decrypted_bytes)
        
        return weights
    
    def aggregate_encrypted(
        self,
        encrypted_weights_list: List[bytes],
        num_samples_list: List[int]
    ) -> List[np.ndarray]:
        """
        Aggregate encrypted weights from multiple clients.
        
        Args:
            encrypted_weights_list: List of encrypted weights from clients
            num_samples_list: List of number of samples per client
            
        Returns:
            Aggregated weights
        """
        # Decrypt all weights
        decrypted_weights_list = []
        for encrypted_weights in encrypted_weights_list:
            weights = self.decrypt_weights(encrypted_weights)
            decrypted_weights_list.append(weights)
        
        # Perform weighted aggregation
        total_samples = sum(num_samples_list)
        
        # Initialize aggregated weights
        aggregated_weights = [np.zeros_like(w) for w in decrypted_weights_list[0]]
        
        # Weighted sum
        for weights, num_samples in zip(decrypted_weights_list, num_samples_list):
            weight_factor = num_samples / total_samples
            for i, w in enumerate(weights):
                aggregated_weights[i] += w * weight_factor
        
        return aggregated_weights


class SimpleSecureAggregation:
    """
    Simplified secure aggregation using additive masking.
    
    This is a basic implementation for demonstration.
    Production systems should use more robust protocols.
    """
    
    def __init__(self, num_clients: int):
        self.num_clients = num_clients
        self.client_masks = {}
        
    def generate_mask(self, client_id: int, shape: tuple) -> np.ndarray:
        """
        Generate random mask for a client.
        
        Args:
            client_id: Client identifier
            shape: Shape of the mask
            
        Returns:
            Random mask array
        """
        np.random.seed(client_id)  # Deterministic for this client
        mask = np.random.randn(*shape)
        self.client_masks[client_id] = mask
        return mask
    
    def mask_weights(
        self,
        weights: List[np.ndarray],
        client_id: int
    ) -> List[np.ndarray]:
        """
        Add mask to weights.
        
        Args:
            weights: Model weights
            client_id: Client identifier
            
        Returns:
            Masked weights
        """
        masked_weights = []
        
        for w in weights:
            mask = self.generate_mask(client_id, w.shape)
            masked_w = w + mask
            masked_weights.append(masked_w)
        
        return masked_weights
    
    def unmask_aggregated(
        self,
        aggregated_weights: List[np.ndarray]
    ) -> List[np.ndarray]:
        """
        Remove masks from aggregated weights.
        
        Args:
            aggregated_weights: Aggregated masked weights
            
        Returns:
            Unmasked weights
        """
        # Sum all masks
        total_masks = [np.zeros_like(w) for w in aggregated_weights]
        
        for client_id in self.client_masks:
            for i, w in enumerate(aggregated_weights):
                mask = self.client_masks[client_id]
                if mask.shape == w.shape:
                    total_masks[i] += mask
        
        # Remove masks
        unmasked_weights = []
        for w, mask in zip(aggregated_weights, total_masks):
            unmasked_w = w - mask
            unmasked_weights.append(unmasked_w)
        
        return unmasked_weights


class HomomorphicAggregator:
    """
    Placeholder for homomorphic encryption-based aggregation.
    
    Note: Full implementation would require libraries like
    TenSEAL or PySEAL for homomorphic encryption.
    """
    
    def __init__(self):
        print("Homomorphic aggregation is a placeholder.")
        print("For production, integrate TenSEAL or similar libraries.")
    
    def encrypt(self, weights: List[np.ndarray]) -> List[np.ndarray]:
        """Placeholder for encryption."""
        # In production, use actual homomorphic encryption
        return weights
    
    def aggregate(
        self,
        encrypted_weights_list: List[List[np.ndarray]],
        num_samples_list: List[int]
    ) -> List[np.ndarray]:
        """Placeholder for homomorphic aggregation."""
        # In production, perform operations on encrypted data
        total_samples = sum(num_samples_list)
        aggregated = [np.zeros_like(w) for w in encrypted_weights_list[0]]
        
        for weights, num_samples in zip(encrypted_weights_list, num_samples_list):
            weight_factor = num_samples / total_samples
            for i, w in enumerate(weights):
                aggregated[i] += w * weight_factor
        
        return aggregated
    
    def decrypt(self, encrypted_weights: List[np.ndarray]) -> List[np.ndarray]:
        """Placeholder for decryption."""
        return encrypted_weights


def aggregate_with_dropout_resilience(
    weights_list: List[List[np.ndarray]],
    num_samples_list: List[int],
    min_clients: int = 2
) -> List[np.ndarray]:
    """
    Aggregate weights with dropout resilience.
    
    Args:
        weights_list: List of weight lists from clients
        num_samples_list: Number of samples per client
        min_clients: Minimum number of clients required
        
    Returns:
        Aggregated weights
    """
    if len(weights_list) < min_clients:
        raise ValueError(f"Insufficient clients: {len(weights_list)} < {min_clients}")
    
    total_samples = sum(num_samples_list)
    aggregated_weights = [np.zeros_like(w) for w in weights_list[0]]
    
    for weights, num_samples in zip(weights_list, num_samples_list):
        weight_factor = num_samples / total_samples
        for i, w in enumerate(weights):
            aggregated_weights[i] += w * weight_factor
    
    return aggregated_weights
