requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Federated Learning
flwr>=1.5.0
flwr-datasets>=0.0.2

# Privacy Preservation
opacus>=1.4.0
pycryptodome>=3.18.0

# Explainability
shap>=0.42.0
lime>=0.2.0.1

# Data Processing
scipy>=1.11.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Dashboard
streamlit>=1.25.0
streamlit-aggrid>=0.3.4

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
